[Darwin] libinotify-kqueue: Add inotify API for BSDs
libinotify-kqueue
Version: 20190118

This is a Darwin/BSD prerequisite for compiling the 'keycheck'
applet incuded with the bootimg utility.

The inotify API and its inherent syscalls are Linux-specific
and thus absent in the *BSD family of operating systems, which
includes Darwin/macOS.

We can work around this problem by including the libinotify-
kqueue library in our build, which provides an inotify API
that uses kqueue(2) to monitor the file system activity.

Requirements for building libinotify-kqueue on Darwin/macOS:

  * Apple's Xcode / Xcode command line tools
  * Homebrew: https://brew.sh
  * gcc, g++, libtool, autoconf, automake

  - Note: gcc, g++ and libtool are included with
    Apple's development tools. However, Apple's
    "gcc" toolchain is not actually based on the
    GNU Compiler Collection, but rather Apple's
    own fork of LLVM/Clang.

    If you run into issues when compiling using
    the default LLVM toolchain, you can try using
    the GNU GCC toolchain provided by Homebrew.

      $ brew install gcc libtool

    As for autoconf and automake, they are not
    present by default. You can install these
    via Homebrew as such:

      $ brew install autoconf automake

Build and install:

  $ autoreconf -fvi
  $ ./configure
  $ make && make install

See libinotify-kqueue/README for further details.

Signed-off-by: Vidir Alexander Jonsson <valex@valex.is>